Make links consistent with button labels
The "Code" button should go to http://imagej.net/Code

And then the Code wiki page can redirect (or not) to the best location
on the wiki for that topic.

Redirects are much better here, to avoid repeated pushing to this Git
repository. By doing things wiki-side, we invite non-developer users to
also update the information linked by those buttons, in an easier way.

The links matching the buttons are also easier to remember and type.