Make main menu context sensitive
Whichever page is open, that section of the menu starts expanded.

Note this requires a new "section:" attribute in the front matter,
for pages that wish to start with a particular section expanded.
But the new update-page-sections.py script can be used to update
all pages to match what's defined in the main-menu include.