Overhaul the page naming
Move menu-linked pages into matching dirs, and downcase the links,
in line with common web conventions.

I used the following script as a starting point:

  cat ../_includes/main-menu | grep -v '^<' |
  sed 's;.*menu-section title="\([^"]*\)" link="\([^"]*\)".*;link=1 \&\& dir=$(echo "\1" | tr "[:upper:]" "[:lower:]") \&\& mkdir "$dir" \&\& cd "$dir" \&\& git mv ~/code/imagej/imagej.github.io/_pages/\2.md ./index.md;' |
  sed 's;.*menu-section .*;link=0;' |
  sed 's;.*menu-section-end.*;test "$link" -eq 1 \&\& cd ..;' |
  sed 's;.*menu-item .* link="\([^"]*\)".*;git mv ~/code/imagej/imagej.github.io/_pages/\1.md ./$(echo "\1" | tr "[:upper:]" "[:lower:]" | tr "_" "-").md;' | sh

And then went through the main menu, tightening up the naming.
The goal is to have URLs which are both short and easy to remember,
with 0-to-2 levels of hierarchy depending on the content in question.