Use "##" instead of "---" for h2 headers
For easier parsing (particularly by the search indexer), I'm going to
change all "------" and "======" style headers to use "##" and "#"
respectively instead. This commit updates only those headers using two
or three single dashes, because they are ambiguous with the Jekyll front
matter, and thus trickier to change programmatically with a script.