Rearrange drag-handle elements
Z index appears to increase with order of declaration, so the "Vital
Statistics" title has actually been cutting off part of the drag-handle.
Since positioning of the drag-handle is absolute it is fine to declare
it after the h3.

position:absolute elements will stick to the nearest parent positioned
element, or the body root if none exists. To be "positioned" just
requires a "position:xxx" style. By making the top statbox div
"position: relative" makes it clear that this is the element we want to
attach absolute items to.