Make the sidebar font a bit larger
It seems it was too small on every screen except for mine.
To accommodate the larger font size, we widen the sidebars,
and shorten the longest menu item titles to stay on one line.

This commit also increases the width cutoff for moving to the
single-column layout. I don't love it; perhaps we should introduce a
two-column intermediate layout between wide screens and narrow screens.