Fix plugin index to use categories array correctly
And change CSS to use flex instead of tables.

There is still a small layout issue when the number of items in a
category modulo 3 is 2, in that the final item becomes centered between
the 2nd and 3rd columns. Surely there is a way to tweak the CSS to fix
this; or if not, we can add a dummy element at the end to round out the
numbers.