Include all categories in the plugin index
Instead of hardcoding, we loop the list of categories, building it up,
and finally sorting and filtering it to unique entries.

We use the Liquid capitalize filter to standardize the category titles
to something reasonable to the human eye, along with a CSS
text-transform rule to capitalize *all* words beyond only the first,
when rendering the actual headers. Unfortunately, this doesn't affect
the TOC rendering, but that's OK -- good enough for now!