Improve the bin scripts
And implement a comprehensive checker for the site.

The individual check-* commands now check one thing each.
And the run-all-checks script runs them all in sequence.
The only exception is check-redirects, which is a temporary
tool intended to exist only until we are done setting up the
Apache 301 redirects for all the old content.