Fix bash-specific _bin scripts to run on Linux
These scripts use the "<<<" construct feeding into a while loop,
to loop arguments with spaces without spawning a subshell.

I prefer not to use bashisms, but I don't know a POSIX-friendly way to
loop over lines with spaces that doesn't spawn a subshell. The construct
"do-a-thing-that-produces-lines-with-spaces | while read line" is nice,
but unfortunately spawns a subshell, and also cannot be nested. We
cannot spawn a subshell in situations where we need to update a variable
in the enclosing scope, because subprocesses cannot touch parent process
state that way. We could write the data to a temp file, then read it
after the subprocess completes, but OMG that's terrible. So let's just
resign ourselves to using a bashism in this situation.