plugins/updater: remove update site mirror info
That document is not only outdated (e.g. we _do_ have rsync daemons
running now), but the setup is likely of interest to a vanishingly small
number of people. All the details detract from the rest of the Updater
page. And anyone wanting to set up a mirror will know to ask on the
Image.sc Forum without there being a lengthy guide on the topic.