Improve the "example" include
And rename it to code-example, since it is specifically a
side-by-side example of raw code, and its formatted result.

And externalize all the CSS into the main CSS rules.

This fixes a problem where the "clear: left" would create undesirable
whitespace after the first code example, if that example occurred prior
to the bottom of the main menu sidebar.