Remove 'template' include
Unclear what it was for. It was only used by the "Fiji's Menu" page,
which has not been updated in many years. It should be possible to
restore this page by enumerating the plugins from Fiji's pom.xml,
should we decide it's necessary. But for now, no point in keeping it.