Add metadata about different project areas
This is the beginning of an attempt to restore the Project template
functionality, so that each page can optionally be part of a project
umbrella, displaying a tooltip over their icon when mousing over the
logo in the statbox, similar to project icons on the old site.

Putting the metadata into the _config.yml makes it extensible, so that
additional projects wanting to be identified on the wiki may do so.