Reconcile help/editing menu with current pages
Now the main menu matches what the nav-links show for these pages.

There is a question of whether we really need all these pages to be
linked in the main menu _and_ with nav-links... but since that's
what we were doing, for now, I just brought it back in sync.

Closes #113.