Speed up main menu generation
We don't need a separate include per menu item. Instead, we can do a
string substitution on the entire menu at the end. We just need to use
<a> elements instead of <b> for the current page (to avoid needing to
replace </a> also with </b>). With corresponding CSS, it works well!

This improves the full build from 38s to 17s on my system.