Move section-specific media into so-named folders
If a media file is only used in e.g. cookbook pages, it now lives in a
corresponding folder media/cookbook. This reduces the number of files
in the base media folder by almost 2000.

This commit also updates the util/find-media include to look first in
the current page's corresponding media folder, so that you don't have
to specify the subfolder in your img include for it to be referenced.

And finally, we add a script to search for media files that are only
used in one specific folder of content, so that we can tidy up the media
locations and references one more time before going to production.