Clean up img include, especially images in tables
With this change, the "image" class is now obsolete.

This also fixes #100 again, after I reintroduced that bug recently.
It makes the max-width-100% images more targeted in two ways:

1. Only img elements from the img include will have the img class.

2. When elements of the img class are inside a table, the max-width is
   reset to 'inherit' to avoid problems with 100% referring to the
   containing table cell rather than the whole content pane.

Admittedly, (2) is a hack and there may be a cleaner way of declaring
that we want 100% of the content pane to be the max width. But I don't
know how, and this avoids the most major downsides of it for now.