Remove Plugins category from all pages
We are only using categories for plugins now. Even we decide
to expand it later, you can tell which wiki pages describe
plugins because they live beneath the /plugins folder.

There are a few plugins that live outside /plugins, such as
the Bio-Formats plugins described in /formats/bio-formats...
but do we really need to use categories to keep track of it?