Simplify the github include
This also changes the default branch from master to main.
Unfortunately, since all the existing links are to older existing
repositories where the default branch is still called master, it means
writing "branch='master'" for all those include usages. But my thinking
is that going forward, as new repositories crop up, and maybe as we
rename the main branches of the existing repositories over time,
we'll benefit from the new default more often. And in the meantime,
this default might bring this issue to people's attention.