Rework the plugin index to have dynamic filtering
And to show all pages with categories, not just plugins.

This is a prototype. It is unoptimized and sluggish.

We don't need to separate pages by category anymore, if we list all the
categories next to each page. And then the filters are more naturally
useful, too. But:

A) Still needs better styling; and
B) Needs bug-fix of "All" condition to actually work.