Move /sites content into /update-sites
Individual plugins are documented in /plugins/*.
And the overview of "what is a plugin" etc. is in /plugins/index.
So analogously, it's OK to mix /sites into /update-sites, especially
since I'm about to change the nav-bar to include nav-links only for
pages which themselves declare nav-links. (So LOCI and IMCF won't
suddenly appear on the nav-bar of the update sites documentation.)

Closes #51, finally.