Remove obsolete categories
Categories are now being used primarily to describe extensions of
ImageJ, for use by the list-of-extensions page. For pages about
software, they are already grouped into /software, so no "Software" or
"Related Software" category is necessary anymore. Similarly, libraries
are in /libs, news posts are in /news, development pages are in
/develop, and so forth.

Perhaps we will reintroduce categories in more context later, but for
the moment, let's limit them to describing types of ImageJ extensions.