Add "Uncategorized" to all pages w/o categories
It seems that 'categories' is a special front matter field in Jekyll,
with every page having this attribute regardless of whether it was
explicitly declared. So we must filter by categories.size > 0. In order
for plugin pages without categories to appear in the list of extensions,
they need to belong to as least one category; the "Uncategorized"
category marks that these plugins require additional classification.