Change /plugin-index to /list-of-extensions
This commit removes the /plugin-index infrastructure, and changes all
links over to /list-of-extensions. It is being done in response to
discussion with the community over the course of this past week.

I am still not certain we want to limit categories to only pages about
extensions (plugins, scripts, update sites). But good enough for now!

fixup! Start on /plugin-index to /list-of-extensions