Make MathJax opt-in via front matter
Trying to auto-detect MathJax syntax in the content only works when that
syntax ($$) is captured. Thus it was accidentally working on the
editing/math page, but broken basically everywhere else.

The generally recommended mechanism for pages to opt-in to MathJax
appears to be via front matter.

Source: https://alanduan.me/random/mathjax/