Use the name ImageJ 1.x instead of ImageJ1
As requested by Wayne Rasband. The name of the original ImageJ is not
"ImageJ1" -- that is just a shorthand many people often use. But the
program is ImageJ, and its version starts with 1., so referring to it as
"ImageJ 1.x" whenever it needs to be distinguished from ImageJ2 is best.