Call the original ImageJ "ImageJ" where possible
And call it "the original ImageJ" instead of "ImageJ 1.x"
when clarification is needed.

This reduces the usages of "ImageJ 1.x" from 391 to 207,
115 of which remain are in archived news posts.

And call ImageJ2 explicitly "ImageJ2" whenever possible.

There are probably still many instances to fix.
It's a wiki, so go ahead and do it if you see them.