Add dygraphs charting library
Later, I'd like to make this an extension. But to do that, we need
to enhance the extension logic to support both header (<head>,
mostly for CSS, but sometimes also JS) and footer (mostly for JS).
Because the dygraphs CSS and JS both need to go in the header, but
the current extensions logic always puts the block in the footer.