Update Update Site Uploads for Github Actions
I duplicated the instructions for Tarvis CI automated uploads to update sites, adapting for GitHub Actions.
The document structure is exactly the same, the main change is the `.yml` configuration file which has to be used.

Please observe however that I had to modify the script that performs the release. In the (older) Tarvis CI section, the bash commands are:

    wget --no-check-certificate https://downloads.imagej.net/fiji/latest/fiji-linux64.zip  &&  unzip fiji-linux64.zip
    mvn clean install -Dscijava.app.directory=$IJ_PATH -Ddelete.other.versions=true
    $IJ_LAUNCHER --update edit-update-site $UPDATE_SITE $URL "webdav:$USER:$WIKI_UPLOAD_PASS" .
    $IJ_LAUNCHER --update update
    $IJ_LAUNCHER --update upload --update-site $UPDATE_SITE --force-shadow jars/YOUR-FILE.jar

whereas with GitHub Actions I suggest (adjusting slightly for readability here):

    curl --silent $IJ_DOWNLOAD_URL | tar --extract --gzip
    $IJ_LAUNCHER --update edit-update-site $UPDATE_SITE $URL "webdav:$USER:$WIKI_UPLOAD_PASS" .
    mvn -B install -Dscijava.app.directory=$IJ_PATH -Ddelete.other.versions=true -Dscijava.ignoreDependencies=true
    $IJ_LAUNCHER --update upload-complete-site --force $UPDATE_SITE

The reason I had to change is that when testing on my side, I figured the steps indicated in the Tarvis CI section create dependency conflicts. I am not sure my fix is adequate, but I would be more than happy to correct it!