Remove package-lock.json from search client
These npm package dependencies are a mess, and the _search/client folder
only affects the actual website tangentially. I don't have time to deal
with the frequent alerts and convoluted version pins that are needed to
address them. Dependabot is very diligent but its fixes can only be as
good as the version constraints imposed by the packages in question. The
most recent PR #197 for example downgrades postcss from 8.2.15 to
7.0.36... really? It's no good. Bye, package-lock.json!

Closes #197.