plugins/image-stitching: update acknowledgements
Remove the link to the ImageJ/non-Fiji version, because the site
no longer exists. Update the list of libraries to link to the
project pages instead of listing JAR file names.

This edit is part of an effort to remove references to the
obsolete and soon-to-be-retired loci_tools.jar bundle.