Statbox: fix source-label autopop
Using liquid's truncate within the if statement seems to create a
problem where the first case is always seen as true (which would set the
source label to "on GitHub" regardless of its value).

If "http(s)://" isn't included the link will resolve as relative so it
was a redundant check. By switching to "includes [domain]" tests we now
also cover when the user has entered leading www.