Fix some ImageJ vs ImageJ2 ambiguities
Specifically: in places where the documentation is ImageJ2-specific,
change ImageJ.app to ImageJ2.app.

As a bonus, because I'm too rushed to exhaustively stage the changes
separately ;-), also eliminate superfluously quoted usages of the bc
template. This template already formats the breadcrumb stylistically,
so including it also in single and/or double quotes is unnecessary.