Vet SIFT plugin page
Unfortunately, the statbox doesn't correctly populate mpicbg:mpicbg_,
because that artifact has an old release versioned 20120621, and even
though I manually hacked all the maven-metadata.xml files I could find
on the server side, and told Nexus to invalidate the cache for
mpicbg:mpicbg_ via its API, the URL:

  https://maven.scijava.org/content/groups/public/mpicbg/mpicbg_/maven-metadata.xml

still serves up a document claiming that 20120621 is the latest version.
Some day, I'll get to the bottom of it, but that day is not today.