Add support for -SNAPSHOT artifacts
It's not a good idea to declare a snapshot artifact as the latest
release. But in some cases, it may be necessary -- for example, the
SLIM-Curve project is discontinued in favor of FLIMJ, but the
slim-curve:slim_plugin component never saw a proper release -- the
latest version available from the SLIM-Curve update site is a snapshot,
and will always be a snapshot. So let's make it possible to note that.