check-include-usage: fix syntax error
The case expression was yielding the following error for me:

  _bin/check-include-usage.sh: line 29: syntax error near unexpected token `;;'
  _bin/check-include-usage.sh: line 29: `    extensions/*) continue ;; # Skip extensions.'

It's strange, because that expression works outside of the while loop.
Maybe it's yet another oddity about "while read" loops, maybe something
to do with subshells, who knows. It is avoidable in this case (ho ho!)
by doing a different equivalent test instead.